reorder graph_access.hpp functions alphabetically

+++ test/test_accessors_dav.cpp
@@ -42,7 +42,7 @@ TEMPLATE_TEST_CASE("dav accessors", "[dav][accessors]", (Graph), (const Graph)) 
 
   vector<double> the_vtx_vals = {10.0, 11.0, 12.0};
 
-  using edge_data_type                = tuple<size_t, size_t, double>;
+  using edge_data_type                = tuple<uint32_t, uint32_t, double>;
   using edge_key_type                 = edge_key_t<G, vertex_edge_iterator_t<G>>;
   vector<edge_data_type> the_edg_vals = {{0, 1, 1.1}, {0, 2, 2.1}, {1, 2, 2.2}, {2, 0, 0.1}}; // {ukey, vkey, val}
